+++ src/game/MSG/MSG.hpp
@@ -2,6 +2,6 @@
 
 #include "../stdafx.hpp"
 
-DWORD* MSG_WriteByte(DWORD* a1, std::int8_t a2);
-DWORD* MSG_WriteBit1(DWORD* a1);
+unsigned long* MSG_WriteByte(unsigned long* a1, std::int8_t a2);
+unsigned long* MSG_WriteBit1(unsigned long* a1);
 void MSG_WriteInt64(DWORD* a1, std::int32_t a2, std::int32_t a3);
